Add soft delete to keep but hide records
This commit adds a 'soft delete' function.  An important thing to note
is that soft deleted records will still be in scope for validations.  In
my mind, that is a needed feature.  If we allowed created records that,
for example, violated uniqueness constraints, then it could pollute the
DB with invalid records.  I would envision notifying the user when that
happens, and then instructing them to refetch with the deleted records.
They could then make the needed changes on that record and submit.  When
the server got those updates, it could restore the record and apply the
update.  There could be an explicit intermediate 'restore' step,
possibly with a simple 'restore' button in the display row that would
then enable the 'Edit' link.  There are certainly other ways to do it,
but you have to really understand the particular use case.  The product
owner's input would be important to make that decision.