Scan for and fix security issues
This commit includes a security report and identifies two issues.  The
first issue is unsafe execution of raw SQL built with unsanitized input.
That issue is fixed in this commit.  The second identified issue is a
potential vulnerability in Rails that may allow forgery of CSRF tokens.
That issue is out of scope for this commit.  The best solution is to
raise the issue with the team and Product Owner so that an upgrade to
the recommended Rails version can be correctly scoped and prioritized.

A third issue that is not in the report, and indeed may be orthogonal,
is the issue of handling 'RecordNotFound.'  Most people would agree that
such errors should be caught and handled gracefully.  OTOH, given that
it should not be easy to select missing records on form inputs, one has
to wonder if the use case is a malicious attacker.  While it may be that
a concurrent user deleted a record in the meantime, i.e. a valid error,
it would be unwise to give too much information to an attacker.  A
discussion with the Product Owner would help decide the requirements
that balance the two concerns.