Perform basic UI cleanup
I didn't leave much time for styling, but I tried to hit the highlights
in terms of making it usable.  Apart from removing the hardcoded field
sizes in favor of applying classes, I would look for a style guideline
or standard for the shop.  In addition to the increased usability that
comes with consistency, I've seen that it's important to sales.  Just
imagine competitors saying, "The problem with their solution is that
it's a Franken-solution.  It's so cobbled together with duct tape,
chewing gum, and bailing wire as a result of all the acquisitions.
You'll see that the pieces don't fit together or work together as they
should, and it's just cumbersome.  I mean, just look at the UI.  It's
different in every section of the product.  Our solution was built from
scratch to work seamlessly together.  Just look at our UI!"

I raised this very issue at Merge/IBM.  We added a card that our UX
person collaborated on.  Sales *Literally* thanked us for it.

If I had more time, I would backfill integration specs.  The tool that
I've used for years is Spinach. https://github.com/codegram/spinach